+++ fcs-core/src/gpu/mod.rs
@@ -4,14 +4,12 @@
 //! batch-norm, activations) as WGSL compute shaders. These building blocks
 //! will power the end-to-end YuNet port in subsequent increments.
 
-pub mod ops;
-#[macro_use]
-pub mod macros;
 pub mod activation;
 pub mod add;
 pub mod batch_norm;
 pub mod conv2d;
 pub mod max_pool;
+pub mod ops;
 pub mod upsample2x;
 pub mod utils;
 
@@ -24,8 +22,7 @@ pub use conv2d::{Conv2dConfig, Conv2dOptions};
 pub use ops::GpuInferenceOps;
 pub mod tensor;
 pub use tensor::{GpuTensor, TensorShape};
-pub mod onnx;
-pub use onnx::{OnnxInitializerMap, OnnxTensor};
+
 const MAX_POOL_WGSL: &str = include_str!("pool.wgsl");
 const ADD_WGSL: &str = include_str!("add.wgsl");
 const UPSAMPLE2X_WGSL: &str = include_str!("resize2x.wgsl");